# [wasm] Add Vector128:ExtractMostSignificantBits