… src/mono/mono/mini/llvm-intrinsics.h
… src/mono/mono/mini/simd-methods.h