@radekdoulik 2022-09-19
@fanyang-mono @matouskozak, arm64 and amd64 implementations of this intrinsic would help on your platforms too I guess